moved saving functionality to own file
